+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Porcupine fever".
 - Tests for "Bijective numeration".
 - Tests for "Plight of the fellowship of the ring".
+- Tests for "DDCG mapper".
 
 ### Fixed
 - Tests groups for "CGFunge interpreter".